requirements.txt
subprocess
time
os
sys
hashlib
pandas
psutil
matplotlib.pyplot
numpy